requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
duckdb>=0.10.0
sentence-transformers>=2.5.0
faiss-cpu>=1.7.4
hdbscan>=0.8.33
umap-learn>=0.5.5
networkx>=3.2
pyvis>=0.3.2
python-louvain>=0.16
google-generativeai>=0.4.0
datamapplot>=0.4.0
scikit-learn>=1.4.0
wikipedia-api>=0.6.0
langdetect>=1.0.9
streamlit-extras>=0.4.0
praw>=7.7.0
